refactor: decompose god files, add DI constructors, migrate to slog
Split oversized files into focused modules:
- logs.go (1906->808 lines): extracted parsing, formatting, streaming, filtering
- checker.go (1077->394 lines): extracted HTTP, command, and process strategies

Add dependency injection constructors:
- wellknown.NewServiceRegistry() for injectable service definitions
- service.NewOperationManager() alongside existing singleton

Clarify executor naming:
- Renamed service/executor.go to service/service_process.go with doc comments
- Added service/doc.go documenting package domains

Extract MCP handler boilerplate:
- handleSingleServiceOp() eliminates repeated validation/controller setup
- Refactored handleStartService and handleRestartService to use helper

Migrate healthcheck logging from zerolog to log/slog:
- checker.go, checker_http.go, monitor.go converted to structured slog calls
- InitializeLogging rewritten using slog handlers

Addresses: #233, #234, #236, #191, #185, #201
Partially addresses: #190 (doc.go added; full split deferred due to 43 importers)
