+++ cli/src/cmd/app/commands/logs.go
@@ -1,16 +1,13 @@
 package commands
 
 import (
-	"bufio"
 	"context"
-	"encoding/json"
 	"fmt"
 	"io"
 	"os"
 	"os/signal"
 	"path/filepath"
 	"strings"
-	"sync"
 	"syscall"
 	"time"
 
@@ -786,144 +783,7 @@ func (e *logsExecutor) collectAzureLogsStandalone(ctx context.Context, cwd strin
 
 // extractLogsWithContext finds log entries matching the level filter and extracts
 // surrounding context lines. Handles deduplication of overlapping context ranges.
-func (e *logsExecutor) extractLogsWithContext(logs []service.LogEntry, levelFilter service.LogLevel, contextLines int) []LogEntryWithContext {
-	if len(logs) == 0 || contextLines <= 0 {
-		return nil
-	}
 
-	// Find indices of matching entries
-	var matchIndices []int
-	for i, entry := range logs {
-		if entry.Level == levelFilter {
-			matchIndices = append(matchIndices, i)
-		}
-	}
-
-	if len(matchIndices) == 0 {
-		return nil
-	}
-
-	// Build entries with context, handling overlapping ranges
-	result := make([]LogEntryWithContext, 0, len(matchIndices))
-	usedIndices := make(map[int]bool) // Track indices already shown in context
-
-	for _, matchIdx := range matchIndices {
-		entry := logs[matchIdx]
-
-		// Extract before context
-		startBefore := matchIdx - contextLines
-		if startBefore < 0 {
-			startBefore = 0
-		}
-		before := make([]string, 0, contextLines)
-		for i := startBefore; i < matchIdx; i++ {
-			// Skip if this line was already shown (to avoid duplicating context)
-			if !usedIndices[i] {
-				before = append(before, logs[i].Message)
-				usedIndices[i] = true
-			}
-		}
-
-		// Extract after context
-		endAfter := matchIdx + contextLines + 1
-		if endAfter > len(logs) {
-			endAfter = len(logs)
-		}
-		after := make([]string, 0, contextLines)
-		for i := matchIdx + 1; i < endAfter; i++ {
-			// Skip if this line was already shown (to avoid duplicating context)
-			if !usedIndices[i] {
-				after = append(after, logs[i].Message)
-				usedIndices[i] = true
-			}
-		}
-
-		// Mark the match itself as used
-		usedIndices[matchIdx] = true
-
-		// Build context only if we have any lines
-		var ctx *LogContext
-		if len(before) > 0 || len(after) > 0 {
-			ctx = &LogContext{
-				Before: before,
-				After:  after,
-			}
-		}
-
-		result = append(result, LogEntryWithContext{
-			Service:   entry.Service,
-			Message:   entry.Message,
-			Level:     logLevelToString(entry.Level),
-			Timestamp: entry.Timestamp,
-			IsStderr:  entry.IsStderr,
-			Context:   ctx,
-		})
-	}
-
-	return result
-}
-
-// logLevelToString converts a LogLevel to its string representation.
-func logLevelToString(level service.LogLevel) string {
-	switch level {
-	case service.LogLevelInfo:
-		return "info"
-	case service.LogLevelWarn:
-		return logLevelWarn
-	case service.LogLevelError:
-		return statusError
-	case service.LogLevelDebug:
-		return logLevelDebug
-	default:
-		return "info"
-	}
-}
-
-// convertAzureLogLevel maps azure.LogLevel to the shared service.LogLevel enum.
-func convertAzureLogLevel(level azure.LogLevel) service.LogLevel {
-	switch level {
-	case azure.LogLevelWarn:
-		return service.LogLevelWarn
-	case azure.LogLevelError:
-		return service.LogLevelError
-	case azure.LogLevelDebug:
-		return service.LogLevelDebug
-	case azure.LogLevelInfo:
-		fallthrough
-	default:
-		return service.LogLevelInfo
-	}
-}
-
-// buildLogFilterInternal creates a log filter from executor options and azure.yaml config.
-func (e *logsExecutor) buildLogFilterInternal(cwd string) (*service.LogFilter, error) {
-	var customPatterns []string
-
-	// Parse command-line exclude patterns
-	if e.opts.exclude != "" {
-		customPatterns = service.ParseExcludePatterns(e.opts.exclude)
-	}
-
-	// Try to load patterns from azure.yaml logs.filters section
-	azureYaml, err := service.ParseAzureYaml(cwd)
-	filterConfig := getFilterConfig(azureYaml, err)
-	if filterConfig != nil {
-		customPatterns = append(customPatterns, filterConfig.Exclude...)
-	}
-
-	// Determine if we should include built-in patterns
-	// CLI --no-builtins flag controls this; azure.yaml always includes builtins
-	includeBuiltins := !e.opts.noBuiltins
-
-	// Build the filter
-	if includeBuiltins {
-		return service.NewLogFilterWithBuiltins(customPatterns)
-	}
-	return service.NewLogFilter(customPatterns)
-}
-
-// getOrCreateSignalChan gets or creates a signal channel with proper cleanup.
-// This avoids duplication and race conditions in signal handling setup.
 func (e *logsExecutor) getOrCreateSignalChan() (chan os.Signal, func()) {
 	if e.signalChan != nil {
 		// Test mode: return existing channel with no-op cleanup
@@ -943,964 +803,3 @@ func (e *logsExecutor) getOrCreateSignalChan() (chan os.Signal, func()) {
 
 // shouldDisplayEntry checks if a log entry should be displayed based on filters.
 // Extracted to avoid code duplication between follow modes.
-func (e *logsExecutor) shouldDisplayEntry(entry service.LogEntry, levelFilter service.LogLevel, logFilter *service.LogFilter) bool {
-	// Filter by level
-	if levelFilter != LogLevelAll && entry.Level != levelFilter {
-		return false
-	}
-
-	// Filter by pattern
-	if logFilter != nil && logFilter.ShouldFilter(entry.Message) {
-		return false
-	}
-
-	return true
-}
-
-// followLogs subscribes to live log streams and displays them.
-func (e *logsExecutor) followLogs(ctx context.Context, projectDir string, logManager LogManagerInterface, dashboardClient DashboardClient, serviceFilter []string, levelFilter service.LogLevel, logFilter *service.LogFilter, outputWriter io.Writer) error {
-	// Handle source-specific follow modes
-	switch e.opts.source {
-	case string(LogSourceAzure):
-		return e.followAzureLogs(ctx, dashboardClient, projectDir, serviceFilter, levelFilter, logFilter, outputWriter)
-	case string(LogSourceAll):
-		return e.followAllLogs(ctx, projectDir, logManager, dashboardClient, serviceFilter, levelFilter, logFilter, outputWriter)
-	default: // "local"
-		return e.followLocalLogs(ctx, projectDir, logManager, dashboardClient, serviceFilter, levelFilter, logFilter, outputWriter)
-	}
-}
-
-// followLocalLogs subscribes to local log streams and displays them.
-func (e *logsExecutor) followLocalLogs(ctx context.Context, _ string, logManager LogManagerInterface, dashboardClient DashboardClient, serviceFilter []string, levelFilter service.LogLevel, logFilter *service.LogFilter, outputWriter io.Writer) error {
-	// Try in-memory subscriptions first
-	subscriptions := make(map[string]chan service.LogEntry)
-
-	if len(serviceFilter) == 0 {
-		// Subscribe to all services
-		for serviceName, buffer := range logManager.GetAllBuffers() {
-			subscriptions[serviceName] = buffer.Subscribe()
-		}
-	} else {
-		// Subscribe to specific services
-		for _, serviceName := range serviceFilter {
-			buffer, exists := logManager.GetBuffer(serviceName)
-			if exists {
-				subscriptions[serviceName] = buffer.Subscribe()
-			}
-		}
-	}
-
-	// If no in-memory buffers, try dashboard WebSocket streaming
-	if len(subscriptions) == 0 {
-		return e.followLogsViaDashboard(ctx, dashboardClient, serviceFilter, levelFilter, logFilter, outputWriter)
-	}
-
-	// Use in-memory streaming
-	return e.followLogsInMemory(subscriptions, logManager, levelFilter, logFilter, outputWriter)
-}
-
-// followLogsViaDashboard connects to the dashboard's WebSocket to stream logs.
-func (e *logsExecutor) followLogsViaDashboard(ctx context.Context, dashboardClient DashboardClient, serviceFilter []string, levelFilter service.LogLevel, logFilter *service.LogFilter, outputWriter io.Writer) error {
-	if dashboardClient == nil {
-		return fmt.Errorf("cannot follow logs: dashboard not available (run 'azd app run' first)")
-	}
-
-	// Check if dashboard is responding
-	if err := dashboardClient.Ping(ctx); err != nil {
-		return fmt.Errorf("cannot follow logs: dashboard not responding (run 'azd app run' first)")
-	}
-
-	cliout.Info("Streaming logs from dashboard...")
-
-	// Create context for streaming that can be canceled
-	streamCtx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	// Setup signal handling for graceful exit
-	sigChan, cleanupSignal := e.getOrCreateSignalChan()
-	defer cleanupSignal()
-
-	// Create channel for log entries
-	logs := make(chan service.LogEntry, logChannelBufferSize)
-
-	// Determine service filter (empty string for all)
-	serviceName := ""
-	if len(serviceFilter) == 1 {
-		serviceName = serviceFilter[0]
-	}
-
-	// Start streaming in background
-	errChan := make(chan error, 1)
-	go func() {
-		errChan <- dashboardClient.StreamLogs(streamCtx, serviceName, logs)
-	}()
-
-	// Display logs as they arrive
-	for {
-		select {
-		case entry := <-logs:
-			// Filter by service if multiple specified
-			if len(serviceFilter) > 1 {
-				found := false
-				for _, svc := range serviceFilter {
-					if entry.Service == svc {
-						found = true
-						break
-					}
-				}
-				if !found {
-					continue
-				}
-			}
-
-			// Use extracted filter method
-			if !e.shouldDisplayEntry(entry, levelFilter, logFilter) {
-				continue
-			}
-
-			// Display log entry
-			if e.opts.format == jsonOutputVal {
-				displayLogsJSON([]service.LogEntry{entry}, outputWriter)
-			} else {
-				displayLogsText([]service.LogEntry{entry}, outputWriter, e.opts.timestamps, e.opts.noColor)
-			}
-
-		case err := <-errChan:
-			if err != nil && err != context.Canceled {
-				return fmt.Errorf("log stream error: %w", err)
-			}
-			return nil
-
-		case <-sigChan:
-			cancel()
-			return nil
-		}
-	}
-}
-
-// followLogsInMemory uses in-memory log buffer subscriptions.
-func (e *logsExecutor) followLogsInMemory(subscriptions map[string]chan service.LogEntry, logManager LogManagerInterface, levelFilter service.LogLevel, logFilter *service.LogFilter, outputWriter io.Writer) error {
-	// Setup signal handling for graceful exit
-	sigChan, cleanupSignal := e.getOrCreateSignalChan()
-	defer cleanupSignal()
-
-	// Create stop channel for goroutine cleanup
-	stopChan := make(chan struct{})
-
-	// Merge all subscription channels with WaitGroup to track completion
-	mergedChan := make(chan service.LogEntry, logChannelBufferSize)
-	var wg sync.WaitGroup
-
-	for _, ch := range subscriptions {
-		wg.Add(1)
-		go func(ch chan service.LogEntry) {
-			defer wg.Done()
-			for {
-				select {
-				case entry, ok := <-ch:
-					if !ok {
-						return
-					}
-					select {
-					case mergedChan <- entry:
-					case <-stopChan:
-						return
-					}
-				case <-stopChan:
-					return
-				}
-			}
-		}(ch)
-	}
-
-	// Close mergedChan when all goroutines complete
-	go func() {
-		wg.Wait()
-		close(mergedChan)
-	}()
-
-	// Cleanup helper function with sync.Once to prevent double-close panics
-	var cleanupOnce sync.Once
-	cleanup := func() {
-		cleanupOnce.Do(func() {
-			close(stopChan)
-			wg.Wait() // Ensure all goroutines stopped before unsubscribing
-			for serviceName, ch := range subscriptions {
-				buffer, exists := logManager.GetBuffer(serviceName)
-				if exists {
-					buffer.Unsubscribe(ch)
-				}
-			}
-		})
-	}
-
-	// Display logs as they arrive
-	for {
-		select {
-		case entry, ok := <-mergedChan:
-			if !ok {
-				// All sources closed
-				cleanup()
-				return nil
-			}
-
-			// Use extracted filter method
-			if !e.shouldDisplayEntry(entry, levelFilter, logFilter) {
-				continue
-			}
-
-			// Display log entry
-			if e.opts.format == jsonOutputVal {
-				displayLogsJSON([]service.LogEntry{entry}, outputWriter)
-			} else {
-				displayLogsText([]service.LogEntry{entry}, outputWriter, e.opts.timestamps, e.opts.noColor)
-			}
-
-		case <-sigChan:
-			cleanup()
-			return nil
-		}
-	}
-}
-
-// followAzureLogs streams Azure logs via the dashboard's WebSocket.
-func (e *logsExecutor) followAzureLogs(ctx context.Context, dashboardClient DashboardClient, projectDir string, serviceFilter []string, levelFilter service.LogLevel, logFilter *service.LogFilter, outputWriter io.Writer) error {
-	// If dashboard is available, prefer its streaming API
-	if dashboardClient != nil {
-		if err := dashboardClient.Ping(ctx); err == nil {
-			status, err := dashboardClient.GetAzureStatus(ctx)
-			if err == nil && status.Enabled {
-				return e.followAzureLogsViaDashboard(ctx, dashboardClient, serviceFilter, levelFilter, logFilter, outputWriter)
-			}
-		}
-	}
-
-	// Fallback: standalone polling (no dashboard needed)
-	return e.followAzureLogsStandalone(ctx, projectDir, serviceFilter, levelFilter, logFilter, outputWriter)
-}
-
-func (e *logsExecutor) followAzureLogsViaDashboard(ctx context.Context, dashboardClient DashboardClient, serviceFilter []string, levelFilter service.LogLevel, logFilter *service.LogFilter, outputWriter io.Writer) error {
-	// Check if dashboard is responding
-	if err := dashboardClient.Ping(ctx); err != nil {
-		return fmt.Errorf("cannot follow Azure logs: dashboard not responding (run 'azd app run' first)")
-	}
-
-	// Check Azure status
-	status, err := dashboardClient.GetAzureStatus(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to get Azure status: %w", err)
-	}
-
-	if !status.Enabled {
-		return fmt.Errorf("azure logging not configured.\n\nTo enable Azure logs:\n  1. Add a Log Analytics configuration to azure.yaml:\n     logs:\n       analytics:\n         pollingInterval: \"30s\"\n         defaultTimespan: \"1h\"\n  2. Ensure your azd environment has workspace outputs (run 'azd provision' if needed)\n  3. Restart 'azd app run'\n\nFor more info: https://aka.ms/azd-app/azure-logs")
-	}
-
-	cliout.Info("Streaming Azure logs (polling every 30s)...")
-
-	// Create context for streaming that can be canceled
-	streamCtx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	// Setup signal handling for graceful exit
-	sigChan, cleanupSignal := e.getOrCreateSignalChan()
-	defer cleanupSignal()
-
-	// Create channel for log entries
-	logs := make(chan service.LogEntry, logChannelBufferSize)
-
-	// Start streaming in background
-	errChan := make(chan error, 1)
-	go func() {
-		errChan <- dashboardClient.StreamAzureLogs(streamCtx, logs)
-	}()
-
-	// Build service filter set if multiple services specified
-	var serviceSet map[string]bool
-	if len(serviceFilter) > 1 {
-		serviceSet = make(map[string]bool)
-		for _, svc := range serviceFilter {
-			serviceSet[svc] = true
-		}
-	}
-
-	// Display logs as they arrive
-	for {
-		select {
-		case entry := <-logs:
-			// Filter by service if multiple specified
-			if serviceSet != nil {
-				if !serviceSet[entry.Service] {
-					continue
-				}
-			} else if len(serviceFilter) == 1 && entry.Service != serviceFilter[0] {
-				continue
-			}
-
-			// Use extracted filter method
-			if !e.shouldDisplayEntry(entry, levelFilter, logFilter) {
-				continue
-			}
-
-			// Display log entry
-			if e.opts.format == jsonOutputVal {
-				displayLogsJSON([]service.LogEntry{entry}, outputWriter)
-			} else {
-				displayLogsText([]service.LogEntry{entry}, outputWriter, e.opts.timestamps, e.opts.noColor)
-			}
-
-		case err := <-errChan:
-			if err != nil && err != context.Canceled {
-				return fmt.Errorf("azure log stream error: %w", err)
-			}
-			return nil
-
-		case <-sigChan:
-			cancel()
-			return nil
-		}
-	}
-}
-
-// followAzureLogsStandalone streams Azure logs without requiring the dashboard.
-func (e *logsExecutor) followAzureLogsStandalone(ctx context.Context, projectDir string, serviceFilter []string, levelFilter service.LogLevel, logFilter *service.LogFilter, outputWriter io.Writer) error {
-	cliout.Info("Streaming Azure logs (standalone, polling)...")
-
-	// Create context for streaming that can be canceled
-	streamCtx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	// Setup signal handling for graceful exit
-	sigChan, cleanupSignal := e.getOrCreateSignalChan()
-	defer cleanupSignal()
-
-	logs := make(chan azure.LogEntry, logChannelBufferSize)
-	errChan := make(chan error, 1)
-
-	initialWindow := time.Hour
-	if e.opts.since != "" {
-		if sinceDuration, err := time.ParseDuration(e.opts.since); err == nil {
-			if sinceDuration < time.Minute {
-				sinceDuration = time.Minute
-			}
-			initialWindow = sinceDuration
-		}
-	}
-
-	pollInterval := 30 * time.Second
-
-	go func() {
-		errChan <- streamAzureLogsStandalone(streamCtx, azure.StreamConfig{
-			ProjectDir:    projectDir,
-			Services:      serviceFilter,
-			PollInterval:  pollInterval,
-			InitialWindow: initialWindow,
-		}, logs)
-	}()
-
-	var serviceSet map[string]bool
-	if len(serviceFilter) > 1 {
-		serviceSet = make(map[string]bool)
-		for _, svc := range serviceFilter {
-			serviceSet[svc] = true
-		}
-	}
-
-	for {
-		select {
-		case azEntry := <-logs:
-			entry := service.LogEntry{
-				Service:   azEntry.Service,
-				Message:   azEntry.Message,
-				Level:     convertAzureLogLevel(azEntry.Level),
-				Timestamp: azEntry.Timestamp,
-				Source:    service.LogSourceAzure,
-				AzureMetadata: &service.AzureLogMetadata{
-					ResourceType:  azEntry.ResourceType,
-					ContainerName: azEntry.ContainerName,
-					InstanceID:    azEntry.InstanceID,
-				},
-			}
-
-			if serviceSet != nil && !serviceSet[entry.Service] {
-				continue
-			}
-			if len(serviceFilter) == 1 && entry.Service != serviceFilter[0] {
-				continue
-			}
-			if !e.shouldDisplayEntry(entry, levelFilter, logFilter) {
-				continue
-			}
-
-			if e.opts.format == jsonOutputVal {
-				displayLogsJSON([]service.LogEntry{entry}, outputWriter)
-			} else {
-				displayLogsText([]service.LogEntry{entry}, outputWriter, e.opts.timestamps, e.opts.noColor)
-			}
-
-		case err := <-errChan:
-			if err != nil && err != context.Canceled {
-				return fmt.Errorf("azure log stream error: %w", err)
-			}
-			return nil
-
-		case <-sigChan:
-			cancel()
-			return nil
-		}
-	}
-}
-
-// followAllLogs streams logs from both local and Azure sources.
-func (e *logsExecutor) followAllLogs(ctx context.Context, projectDir string, _ LogManagerInterface, dashboardClient DashboardClient, serviceFilter []string, levelFilter service.LogLevel, logFilter *service.LogFilter, outputWriter io.Writer) error {
-	if dashboardClient == nil {
-		cliout.Warning("Dashboard not running; following Azure logs only.")
-		return e.followAzureLogsStandalone(ctx, projectDir, serviceFilter, levelFilter, logFilter, outputWriter)
-	}
-
-	// Check if dashboard is responding
-	if err := dashboardClient.Ping(ctx); err != nil {
-		cliout.Warning("Dashboard not responding; following Azure logs only.")
-		return e.followAzureLogsStandalone(ctx, projectDir, serviceFilter, levelFilter, logFilter, outputWriter)
-	}
-
-	cliout.Info("Streaming logs from local and Azure sources...")
-
-	// Create context for streaming that can be canceled
-	streamCtx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	// Setup signal handling for graceful exit
-	sigChan, cleanupSignal := e.getOrCreateSignalChan()
-	defer cleanupSignal()
-
-	// Create merged channel for all log entries
-	mergedLogs := make(chan service.LogEntry, logChannelBufferSize*2)
-
-	// Start local log streaming
-	localLogs := make(chan service.LogEntry, logChannelBufferSize)
-	serviceName := ""
-	if len(serviceFilter) == 1 {
-		serviceName = serviceFilter[0]
-	}
-
-	var wg sync.WaitGroup
-	errChan := make(chan error, 2)
-
-	// Local logs goroutine
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		defer close(localLogs)
-		if err := dashboardClient.StreamLogs(streamCtx, serviceName, localLogs); err != nil && err != context.Canceled {
-			errChan <- fmt.Errorf("local log stream error: %w", err)
-		}
-	}()
-
-	// Azure logs goroutine (only if Azure is enabled)
-	azureLogs := make(chan service.LogEntry, logChannelBufferSize)
-	status, _ := dashboardClient.GetAzureStatus(ctx)
-	if status != nil && status.Enabled {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			defer close(azureLogs)
-			if err := dashboardClient.StreamAzureLogs(streamCtx, azureLogs); err != nil && err != context.Canceled {
-				// Azure errors are non-fatal, just log and continue
-				cliout.Warning("Azure log stream disconnected: %s", err)
-			}
-		}()
-	} else {
-		close(azureLogs)
-	}
-
-	// Merger goroutine - combines local and azure logs
-	go func() {
-		for {
-			select {
-			case entry, ok := <-localLogs:
-				if ok {
-					mergedLogs <- entry
-				}
-			case entry, ok := <-azureLogs:
-				if ok {
-					mergedLogs <- entry
-				}
-			case <-streamCtx.Done():
-				return
-			}
-		}
-	}()
-
-	// Build service filter set if multiple services specified
-	var serviceSet map[string]bool
-	if len(serviceFilter) > 1 {
-		serviceSet = make(map[string]bool)
-		for _, svc := range serviceFilter {
-			serviceSet[svc] = true
-		}
-	}
-
-	// Display logs as they arrive
-	for {
-		select {
-		case entry := <-mergedLogs:
-			// Filter by service if multiple specified
-			if serviceSet != nil && !serviceSet[entry.Service] {
-				continue
-			}
-
-			// Use extracted filter method
-			if !e.shouldDisplayEntry(entry, levelFilter, logFilter) {
-				continue
-			}
-
-			// Display log entry
-			if e.opts.format == jsonOutputVal {
-				displayLogsJSON([]service.LogEntry{entry}, outputWriter)
-			} else {
-				displayLogsText([]service.LogEntry{entry}, outputWriter, e.opts.timestamps, e.opts.noColor)
-			}
-
-		case err := <-errChan:
-			if err != nil {
-				return err
-			}
-
-		case <-sigChan:
-			cancel()
-			return nil
-		}
-	}
-}
-
-// readLogsFromFile reads logs from the persisted log file for a service.
-// This is used when the in-memory buffer is empty (e.g., when called from a subprocess).
-// It also reads from rotated backup files (.log.1, .log.2) if needed.
-func readLogsFromFile(projectDir, serviceName string, tail int, sinceTime time.Time) ([]service.LogEntry, error) {
-	logsDir := filepath.Join(projectDir, ".azure", "logs")
-	baseLogFile := filepath.Join(logsDir, serviceName+".log")
-
-	var allEntries []service.LogEntry
-
-	// Read from rotated files first (oldest to newest: .log.2, .log.1, .log)
-	logFiles := []string{
-		baseLogFile + ".2",
-		baseLogFile + ".1",
-		baseLogFile,
-	}
-
-	for _, logFile := range logFiles {
-		entries, err := readSingleLogFile(logFile, serviceName, sinceTime)
-		if err != nil {
-			continue // File may not exist (rotated files are optional)
-		}
-		allEntries = append(allEntries, entries...)
-	}
-
-	if len(allEntries) == 0 {
-		return nil, fmt.Errorf("no log files found for service %s", serviceName)
-	}
-
-	// Apply tail limit
-	if tail > 0 && len(allEntries) > tail {
-		allEntries = allEntries[len(allEntries)-tail:]
-	}
-
-	return allEntries, nil
-}
-
-// readSingleLogFile reads log entries from a single log file.
-func readSingleLogFile(logFile, serviceName string, sinceTime time.Time) ([]service.LogEntry, error) {
-	cleanLogFile := filepath.Clean(logFile)
-	file, err := os.Open(cleanLogFile)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = file.Close() }()
-
-	var entries []service.LogEntry
-	scanner := bufio.NewScanner(file)
-	// Increase buffer size to handle long log lines (stack traces, JSON dumps)
-	scanner.Buffer(make([]byte, scannerInitialBufferSize), maxLogLineSize)
-
-	for scanner.Scan() {
-		line := scanner.Text()
-		entry, err := parseLogLine(line, serviceName)
-		if err != nil {
-			continue // Skip unparseable lines
-		}
-
-		// Apply since filter
-		if !sinceTime.IsZero() && entry.Timestamp.Before(sinceTime) {
-			continue
-		}
-
-		entries = append(entries, entry)
-	}
-
-	if err := scanner.Err(); err != nil {
-		return nil, err
-	}
-
-	return entries, nil
-}
-
-// parseLogLine parses a log line from the file format:
-// [2006-01-02 15:04:05.000] [LEVEL] [STREAM] message
-func parseLogLine(line, serviceName string) (service.LogEntry, error) {
-	entry := service.LogEntry{
-		Service: serviceName,
-	}
-
-	// Parse timestamp: [2006-01-02 15:04:05.000]
-	if len(line) < 25 || line[0] != '[' {
-		return entry, fmt.Errorf("invalid log line format")
-	}
-
-	endTimestamp := strings.Index(line[1:], "]")
-	if endTimestamp == -1 {
-		return entry, fmt.Errorf("missing timestamp end bracket")
-	}
-
-	timestampStr := line[1 : endTimestamp+1]
-	timestamp, err := time.Parse("2006-01-02 15:04:05.000", timestampStr)
-	if err != nil {
-		return entry, fmt.Errorf("failed to parse timestamp: %w", err)
-	}
-	entry.Timestamp = timestamp
-
-	// Parse remaining: [LEVEL] [STREAM] message
-	remaining := line[endTimestamp+3:] // Skip "] "
-
-	// Parse level: [LEVEL]
-	if len(remaining) < 3 || remaining[0] != '[' {
-		entry.Message = remaining
-		entry.Level = service.LogLevelInfo
-		return entry, nil
-	}
-
-	endLevel := strings.Index(remaining[1:], "]")
-	if endLevel == -1 {
-		entry.Message = remaining
-		entry.Level = service.LogLevelInfo
-		return entry, nil
-	}
-
-	levelStr := remaining[1 : endLevel+1]
-	entry.Level = parseLogLevelFromString(levelStr)
-	remaining = remaining[endLevel+3:] // Skip "] "
-
-	// Parse stream: [STREAM]
-	if len(remaining) >= 3 && remaining[0] == '[' {
-		endStream := strings.Index(remaining[1:], "]")
-		if endStream != -1 {
-			streamStr := remaining[1 : endStream+1]
-			entry.IsStderr = streamStr == "ERR"
-			remaining = remaining[endStream+3:] // Skip "] "
-		}
-	}
-
-	entry.Message = remaining
-	return entry, nil
-}
-
-// parseLogLevelFromString parses a log level from a string.
-func parseLogLevelFromString(level string) service.LogLevel {
-	switch strings.ToUpper(level) {
-	case "INFO":
-		return service.LogLevelInfo
-	case "WARN", "WARNING":
-		return service.LogLevelWarn
-	case "ERROR":
-		return service.LogLevelError
-	case "DEBUG":
-		return service.LogLevelDebug
-	default:
-		return service.LogLevelInfo
-	}
-}
-
-// ANSI color constants for log output formatting.
-// colorCyan is defined here as it's not in info.go.
-// Other colors (colorGray, colorRed, colorYellow, colorReset) are in info.go.
-const colorCyan = "\033[36m"
-
-// displayLogsText displays logs in text format.
-// Uses io.Writer interface for better testability and flexibility.
-func displayLogsText(logs []service.LogEntry, w io.Writer, showTimestamps, noColor bool) {
-	for _, entry := range logs {
-		var line strings.Builder
-
-		// Timestamp
-		if showTimestamps {
-			timestamp := entry.Timestamp.Format("15:04:05.000")
-			if noColor {
-				fmt.Fprintf(&line, "[%s] ", timestamp)
-			} else {
-				line.WriteString(colorGray + "[" + timestamp + "]" + colorReset + " ")
-			}
-		}
-
-		// Service name
-		if noColor {
-			fmt.Fprintf(&line, "[%s] ", entry.Service)
-		} else {
-			line.WriteString(colorCyan + "[" + entry.Service + "]" + colorReset + " ")
-		}
-
-		// Message with color based on stderr/level
-		if noColor {
-			line.WriteString(entry.Message)
-		} else {
-			if entry.IsStderr || entry.Level == service.LogLevelError {
-				line.WriteString(colorRed + entry.Message + colorReset)
-			} else if entry.Level == service.LogLevelWarn {
-				line.WriteString(colorYellow + entry.Message + colorReset)
-			} else if entry.Level == service.LogLevelDebug {
-				line.WriteString(colorGray + entry.Message + colorReset)
-			} else {
-				line.WriteString(entry.Message)
-			}
-		}
-
-		_, _ = fmt.Fprintln(w, line.String())
-	}
-}
-
-// displayLogsJSON displays logs in JSON format.
-// Uses io.Writer interface for better testability and flexibility.
-func displayLogsJSON(logs []service.LogEntry, w io.Writer) {
-	encoder := json.NewEncoder(w)
-	for _, entry := range logs {
-		if err := encoder.Encode(entry); err != nil {
-			fmt.Fprintf(os.Stderr, "Warning: failed to encode log entry: %v\n", err)
-		}
-	}
-}
-
-// displayLogsWithContextJSON displays logs with context in JSON format.
-// Each entry includes optional before/after context lines.
-func displayLogsWithContextJSON(logs []LogEntryWithContext, w io.Writer) {
-	encoder := json.NewEncoder(w)
-	for _, entry := range logs {
-		if err := encoder.Encode(entry); err != nil {
-			fmt.Fprintf(os.Stderr, "Warning: failed to encode log entry: %v\n", err)
-		}
-	}
-}
-
-// displayLogsWithContextText displays logs with context in text format.
-// Context lines are shown with indentation and separators between entries.
-func displayLogsWithContextText(logs []LogEntryWithContext, w io.Writer, showTimestamps, noColor bool) {
-	for i, entry := range logs {
-		// Add separator between entries (not before first)
-		if i > 0 {
-			_, _ = fmt.Fprintln(w, "---")
-		}
-
-		// Show before context (if any)
-		if entry.Context != nil && len(entry.Context.Before) > 0 {
-			for _, line := range entry.Context.Before {
-				if noColor {
-					_, _ = fmt.Fprintf(w, "  %s\n", line)
-				} else {
-					_, _ = fmt.Fprintf(w, "  %s%s%s\n", colorGray, line, colorReset)
-				}
-			}
-		}
-
-		// Show the matching entry
-		var line strings.Builder
-
-		// Timestamp
-		if showTimestamps {
-			timestamp := entry.Timestamp.Format("15:04:05.000")
-			if noColor {
-				fmt.Fprintf(&line, "[%s] ", timestamp)
-			} else {
-				line.WriteString(colorGray + "[" + timestamp + "]" + colorReset + " ")
-			}
-		}
-
-		// Service name
-		if noColor {
-			fmt.Fprintf(&line, "[%s] ", entry.Service)
-		} else {
-			line.WriteString(colorCyan + "[" + entry.Service + "]" + colorReset + " ")
-		}
-
-		// Message with color based on level
-		if noColor {
-			line.WriteString(entry.Message)
-		} else {
-			switch entry.Level {
-			case "error":
-				line.WriteString(colorRed + entry.Message + colorReset)
-			case logLevelWarn:
-				line.WriteString(colorYellow + entry.Message + colorReset)
-			case logLevelDebug:
-				line.WriteString(colorGray + entry.Message + colorReset)
-			default:
-				line.WriteString(entry.Message)
-			}
-		}
-
-		_, _ = fmt.Fprintln(w, line.String())
-
-		// Show after context (if any)
-		if entry.Context != nil && len(entry.Context.After) > 0 {
-			for _, contextLine := range entry.Context.After {
-				if noColor {
-					_, _ = fmt.Fprintf(w, "  %s\n", contextLine)
-				} else {
-					_, _ = fmt.Fprintf(w, "  %s%s%s\n", colorGray, contextLine, colorReset)
-				}
-			}
-		}
-	}
-}
-
-// LogLevelAll is a sentinel value indicating no level filtering should be applied.
-const LogLevelAll service.LogLevel = -1
-
-// parseLogLevel parses a log level string.
-func parseLogLevel(level string) service.LogLevel {
-	switch strings.ToLower(level) {
-	case "info":
-		return service.LogLevelInfo
-	case logLevelWarn, "warning":
-		return service.LogLevelWarn
-	case "error":
-		return service.LogLevelError
-	case logLevelDebug:
-		return service.LogLevelDebug
-	case "all":
-		return LogLevelAll
-	default:
-		return LogLevelAll
-	}
-}
-
-// filterLogsByLevel filters logs by level with pre-allocated capacity.
-func filterLogsByLevel(logs []service.LogEntry, level service.LogLevel) []service.LogEntry {
-	if level == LogLevelAll {
-		return logs
-	}
-
-	// Pre-allocate with estimated capacity based on typical match rate
-	estimatedCap := len(logs) / filterCapacityEstimate
-	if estimatedCap < 10 {
-		estimatedCap = 10
-	}
-	filtered := make([]service.LogEntry, 0, estimatedCap)
-	for _, entry := range logs {
-		if entry.Level == level {
-			filtered = append(filtered, entry)
-		}
-	}
-	return filtered
-}
-
-// buildLogFilter creates a log filter from options and azure.yaml config.
-// This is a test helper function that wraps buildLogFilterInternal.
-//
-// Deprecated: Use executor.buildLogFilterInternal directly in new code.
-func buildLogFilter(cwd string, exclude string, noBuiltins bool) (*service.LogFilter, error) {
-	var customPatterns []string
-
-	// Parse command-line exclude patterns
-	if exclude != "" {
-		customPatterns = service.ParseExcludePatterns(exclude)
-	}
-
-	// Try to load patterns from azure.yaml logs.filters section
-	azureYaml, err := service.ParseAzureYaml(cwd)
-	filterConfig := getFilterConfig(azureYaml, err)
-	if filterConfig != nil {
-		customPatterns = append(customPatterns, filterConfig.Exclude...)
-	}
-
-	// Determine if we should include built-in patterns
-	// CLI --no-builtins flag controls this; azure.yaml always includes builtins
-	includeBuiltins := !noBuiltins
-
-	// Build the filter
-	if includeBuiltins {
-		return service.NewLogFilterWithBuiltins(customPatterns)
-	}
-	return service.NewLogFilter(customPatterns)
-}
-
-// getFilterConfig extracts the filter config from azure.yaml's logs section.
-func getFilterConfig(azureYaml *service.AzureYaml, err error) *service.LogFilterConfig {
-	if err != nil || azureYaml == nil {
-		return nil
-	}
-	return azureYaml.Logs.GetFilters()
-}
-
-// validateLogsOptions validates command-line flag values.
-func validateLogsOptions(opts *logsOptions) error {
-	// Validate tail is positive
-	if opts.tail < 0 {
-		return fmt.Errorf("--tail must be a positive number, got %d", opts.tail)
-	}
-	if opts.tail > maxTailLines {
-		// Log warning before capping
-		fmt.Fprintf(os.Stderr, "Warning: --tail value %d exceeds maximum, capping at %d\n", opts.tail, maxTailLines)
-		opts.tail = maxTailLines
-	}
-
-	// Validate format
-	switch opts.format {
-	case "text", jsonOutputVal:
-		// Valid formats
-	default:
-		return fmt.Errorf("--format must be 'text' or 'json', got '%s'", opts.format)
-	}
-
-	// Validate level
-	switch strings.ToLower(opts.level) {
-	case "info", logLevelWarn, "warning", "error", logLevelDebug, "all":
-		// Valid levels
-	default:
-		return fmt.Errorf("--level must be one of: info, warn, error, debug, all; got '%s'", opts.level)
-	}
-
-	// Validate context requires level to be set (not "all")
-	if opts.contextLines > 0 {
-		if strings.ToLower(opts.level) == "all" {
-			return fmt.Errorf("--context requires --level to be set (info, warn, error, or debug)")
-		}
-	}
-
-	// Clamp context to valid range (0-MaxContextLines)
-	if opts.contextLines < 0 {
-		opts.contextLines = 0
-	}
-	if opts.contextLines > service.MaxContextLines {
-		fmt.Fprintf(os.Stderr, "Warning: --context value %d exceeds maximum, capping at %d\n", opts.contextLines, service.MaxContextLines)
-		opts.contextLines = service.MaxContextLines
-	}
-
-	// Validate since duration if provided
-	if opts.since != "" {
-		if _, err := time.ParseDuration(opts.since); err != nil {
-			return fmt.Errorf("--since must be a valid duration (e.g., 5m, 1h), got '%s': %w", opts.since, err)
-		}
-	}
-
-	// Validate source
-	switch strings.ToLower(opts.source) {
-	case string(LogSourceLocal), string(LogSourceAzure), string(LogSourceAll):
-		// Valid sources - normalize to lowercase
-		opts.source = strings.ToLower(opts.source)
-	case "":
-		// Default to local if not specified
-		opts.source = string(LogSourceLocal)
-	default:
-		return fmt.Errorf("--source must be 'local', 'azure', or 'all'; got '%s'", opts.source)
-	}
-
-	return nil
-}
